requirements.txt
torch>=2.2
torchvision>=0.17
pillow>=10.3
numpy>=1.26
opencv-python>=4.8
